+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - CodiMD instances can now be branded either with a '@ <custom string>' or '@ <custom logo>' after the CodiMD logo and text
 - Images will be loaded via proxy if an image proxy is configured in the backend
 - Asciinema videos may now be embedded by pasting the URL of one video into a single line
+- The Toolbar includes an EmojiPicker
 
 ### Changed